refactor: Split cli.rs into cli/ module structure
- Convert cli.rs to cli/mod.rs (~1814 lines)
- Extract display functions to cli/display.rs (~191 lines)
  - read_context_lines
  - display_unified_results
  - display_unified_results_json

This is the first step of splitting the ~2000 line cli.rs. Further
extraction of commands, pipeline, and utility functions can follow.

Partial fix for #78
